Configure examples to play nice together in the same directory
